requirements.txt
uagents==0.24
openai
supabase
python-dotenv
